+++ client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # PDDL support - What's new?
 
+## [2.15.3] Bug fixes
+
+- Fixed generated problem file/tab name for the pre-parse preview
+- Support for spaces in problem template pre-process command-line
+- Demoted Jinja2 errors to parsing problems make it possible to edit the files
+- Configuration selector menu shows current value
+- Introduced configuration to disable the model hierarchy decorator (for perf reasons)
+- Added PDDL3.1 `undefined` to syntax highlighting
+- To address the [Issue #49](https://github.com/jan-dolejsi/vscode-pddl/issues/49) confusion why the search debugger is not displaying anything, when a planning service is called, the extension will now show a warning message that the search debugger config switch is being ignored.
+- Removed tips that were not effective at teaching how to use the more hidden features [Issue #47](https://github.com/jan-dolejsi/vscode-pddl/issues/49) as suggested and fixed by [@boramalper](https://github.com/jan-dolejsi/vscode-pddl/pull/48)
+
 ## [2.15.2] Predicate and function references
 
 - predicate/function references in hover and decorations (image needed)
@@ -865,7 +876,8 @@ Note for open source contributors: all notable changes to the "pddl" extension w
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-[Unreleased]: https://github.com/jan-dolejsi/vscode-pddl/compare/v2.15.2...HEAD
+[Unreleased]: https://github.com/jan-dolejsi/vscode-pddl/compare/v2.15.3...HEAD
+[2.15.3]:https://github.com/jan-dolejsi/vscode-pddl/compare/v2.15.2...v2.15.3
 [2.15.2]:https://github.com/jan-dolejsi/vscode-pddl/compare/v2.15.1...v2.15.2
 [2.15.1]:https://github.com/jan-dolejsi/vscode-pddl/compare/v2.15.0...v2.15.1
 [2.15.0]:https://github.com/jan-dolejsi/vscode-pddl/compare/v2.14.4...v2.15.0